trylytix_backend/requirements.txt
asgiref==3.8.1
awsebcli==3.23.3
blessed==1.21.0
botocore==1.38.14
cement==2.10.14
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
joblib==1.5.0
numpy==2.2.5
nvidia-nccl-cu12==2.26.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
setuptools==80.4.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
xgboost==3.0.1
